Review changes

+++ Source/Server/Cuelogic.Clrm.Api/Controllers/AllocationController.cs
@@ -25,6 +25,8 @@ namespace Cuelogic.Clrm.Api.Controllers
         [Route("")]
         public IHttpActionResult Get(int show, int page, string filterText)
         {
+            if (show < 0 || page < 0)
+                throw new Exception("Negative values not allowed");
             var searchParam = new SearchParam();
             searchParam.FilterText = filterText ?? "";
             searchParam.Page = page;
@@ -37,6 +39,8 @@ namespace Cuelogic.Clrm.Api.Controllers
         [Route("{id}")]
         public IHttpActionResult Get(int id)
         {
+            if (id < 0)
+                throw new Exception("Negative id now allowed");
             var allocation = _allocationService.GetItem(id);
             return Ok(allocation);
         }
@@ -45,6 +49,8 @@ namespace Cuelogic.Clrm.Api.Controllers
         [Route("GetAllocation/{id}")]
         public IHttpActionResult GetAllocation(int id)
         {
+            if (id < 0)
+                throw new Exception("Negative id now allowed");
             var allocation = _allocationService.GetAllocationSum(id);
             return Ok(allocation);
         }
@@ -62,6 +68,8 @@ namespace Cuelogic.Clrm.Api.Controllers
         [Route("{id}")]
         public IHttpActionResult Delete(int id)
         {
+            if (id < 0)
+                throw new Exception("Negative id now allowed");
             _allocationService.Delete(id);
             return Ok();
         }
